+++ test/libsolidity/syntaxTests/immutable/non-value_type_struct.sol
@@ -1,0 +1,9 @@
+contract Contract {
+    struct S {
+        int k;
+    }
+
+    S immutable s;
+}
+// ----
+// TypeError 6377: (61-74): Immutable variables cannot have a non-value type.