+++ test/libsolidity/syntaxTests/immutable/non-value_type_struct_with_mapping.sol
@@ -1,0 +1,9 @@
+contract Contract {
+    struct S {
+        mapping(uint => address) map;
+    }
+
+    S immutable s;
+}
+// ----
+// TypeError 6377: (84-97): Immutable variables cannot have a non-value type.